refactor(rust): break down methods in link_stage/mod.rs into individual files
