+++ crates/rolldown/src/stages/link_stage/generate_lazy_export.rs
@@ -21,7 +21,7 @@ use rolldown_utils::{
 use super::LinkStage;
 
 impl LinkStage<'_> {
-  pub fn generate_lazy_export(&mut self) {
+  pub(super) fn generate_lazy_export(&mut self) {
     let module_idx_to_exports_kind = append_only_vec::AppendOnlyVec::new();
     self.module_table.modules.par_iter_mut().for_each(|module| {
       let Module::Normal(module) = module else {
