+++ crates/rolldown/src/stages/link_stage/wrapping.rs
@@ -75,7 +75,7 @@ fn has_dynamic_exports_due_to_export_star(
 
 impl LinkStage<'_> {
   #[tracing::instrument(level = "debug", skip_all)]
-  pub fn wrap_modules(&mut self) {
+  pub(super) fn wrap_modules(&mut self) {
     let mut visited_modules_for_wrapping =
       oxc_index::index_vec![false; self.module_table.modules.len()];
     let mut visited_modules_for_dynamic_exports =
